+++ server/core/runtime/run_step_runner.go
@@ -25,11 +25,13 @@ func (r *RunStepRunner) Run(ctx models.ProjectCommandContext, command string, pa
 		tfVersion = ctx.TerraformVersion
 	}
 
-	err := r.TerraformExecutor.EnsureVersion(ctx.Log, tfVersion)
-	if err != nil {
-		err = fmt.Errorf("%s: Downloading terraform Version %s", err, tfVersion.String())
-		ctx.Log.Debug("error: %s", err)
-		return "", err
+	if ctx.TerraformVersion != nil {
+		err := r.TerraformExecutor.EnsureVersion(ctx.Log, tfVersion)
+		if err != nil {
+			err = fmt.Errorf("%s: Downloading terraform Version %s", err, tfVersion.String())
+			ctx.Log.Debug("error: %s", err)
+			return "", err
+		}
 	}
 
 	cmd := exec.Command("sh", "-c", command) // #nosec
